+++ scripts/make/data-fetcher.lua
@@ -1,26 +1,29 @@
 local toml = require 'toml'
 local Logs = require 'lib.logs'
+local i18n = require 'lib.i18n'
+local fun = require 'lib.fun'
 
 local DataFetcher = {}
+
+local function errmsg(err)
+  if type(err) ~= 'string' then return end
+  return err:match('TOML:%s*(.*)$')
+end
 
 local function read_file(fp)
   local f, msg = io.open(fp, 'r')
   Logs.assert(f, msg)
   local str = f:read('*a')
   f:close()
-  return str
+  local s, p = pcall(toml.parse, str)
+  Logs.assert(s, i18n 'make.data_fetcher.toml_error', ' ', errmsg(p))
+  return p
 end
 
-local function merge(files)
-  local raw = ''
-  for _, fp in ipairs(files) do
-    raw = raw .. read_file(fp) .. '\n'
-  end
-  return toml.parse(raw)
-end
-
+--- @param files Fun
+--- @return Fun
 function DataFetcher.get(files)
-  return merge(files)
+  return fun {}:merge(unpack(files:map(read_file)))
 end
 
 return DataFetcher