+++ scripts/ygofab.lua
@@ -69,7 +69,7 @@ interpreter:add_command('compose', cmd_compose, '-p', 1, '-Pall', 0, '-e', 1,
 interpreter:add_command('config', cmd_config)
 interpreter:add_command('export', cmd_export, '-p', 1, '-Pall', 0, '-o', 1,
                         '-e', 1, '-Eall', 0, '--verbose', 0)
-interpreter:add_command('make', cmd_make, '--clean', 0, '--all', 0)
+interpreter:add_command('make', cmd_make, '--overwrite', 0, '-ow', 0, '--all', 0)
 interpreter:add_command('new', cmd_new)
 interpreter:add_command('sync', cmd_sync, '-g', 1, '-Gall', 0, '-p', 1, '-e', 1,
                         '--no-string', 0, '--verbose', 0)